fix parse.one_of for cases where parsing stops early on partial matches
  - add tests
  - add better type annotations

Change to parse API: formatter parameter is no longer auto-curried.
Currying was not working anyway for the most common case, a class constructor.
So making it explicit: if you are parsing > 1 element from the request, you
must supply a curried function. This is made easier now with util.adt.Type.

minor bug fixes in util.adt